+++ src/engine/search/principal_variation.rs
@@ -1,55 +1,54 @@
 use crate::chess::moves::Move;
-use crate::chess::square::squares::all::*;
 use crate::engine::search::MAX_SEARCH_DEPTH_SIZE;
+use arrayvec::ArrayVec;
 
 #[derive(Clone)]
-pub struct PrincipalVariation {
-    moves: [Move; MAX_SEARCH_DEPTH_SIZE],
-    length: usize,
-}
+pub struct PrincipalVariation(ArrayVec<Move, MAX_SEARCH_DEPTH_SIZE>);
 
 impl PrincipalVariation {
-    const EMPTY_MOVE: Move = Move::new(A1, A1);
-
+    #[inline]
     pub const fn new() -> Self {
-        Self {
-            moves: [Self::EMPTY_MOVE; MAX_SEARCH_DEPTH_SIZE],
-            length: 0,
-        }
-    }
-
-    pub fn clear(&mut self) {
-        self.length = 0;
-    }
-
-    pub fn push(&mut self, mv: Move, child_pv: &Self) {
-        self.length = child_pv.length + 1;
-        self.moves[0] = mv;
-        self.moves[1..self.length].copy_from_slice(child_pv.as_slice());
+        Self(ArrayVec::new_const())
     }
 
     #[inline]
-    pub fn as_slice(&self) -> &[Move] {
-        &self.moves[0..self.length]
+    pub fn clear(&mut self) {
+        self.0.clear();
     }
 
-    pub fn first(&self) -> Option<Move> {
-        if self.length == 0 {
-            None
-        } else {
-            Some(self.moves[0])
-        }
+    #[inline]
+    pub fn push(&mut self, mv: Move, child_pv: &Self) {
+        self.0.clear();
+        self.0.push(mv);
+        self.0
+            .try_extend_from_slice(&child_pv.0)
+            .expect("Could not construct PV");
+    }
+
+    #[inline]
+    pub fn first(&self) -> Option<&Move> {
+        self.0.first()
     }
 
     #[cfg(test)]
     pub fn len(&self) -> usize {
-        self.length
+        self.0.len()
+    }
+}
+
+impl IntoIterator for PrincipalVariation {
+    type Item = Move;
+    type IntoIter = arrayvec::IntoIter<Self::Item, MAX_SEARCH_DEPTH_SIZE>;
+
+    fn into_iter(self) -> Self::IntoIter {
+        self.0.into_iter()
     }
 }
 
 #[cfg(test)]
 mod tests {
     use super::*;
+    use crate::chess::square::squares::all::*;
 
     #[test]
     fn test_new_pv_has_no_moves() {
@@ -71,8 +70,8 @@ mod tests {
         pv_4.push(Move::new(E1, F1), &pv_3);
 
         assert_eq!(pv_4.len(), 3);
-        assert_eq!(pv_4.moves[0].src, E1);
-        assert_eq!(pv_4.moves[1].src, C1);
-        assert_eq!(pv_4.moves[2].src, A1);
+        assert_eq!(pv_4.0.get(0).unwrap().src, E1);
+        assert_eq!(pv_4.0.get(1).unwrap().src, C1);
+        assert_eq!(pv_4.0.get(2).unwrap().src, A1);
     }
 }