+++ src/engine/uci/mod.rs
@@ -53,7 +53,7 @@ impl UciReporter {
             depth: Some(progress.depth),
             seldepth: Some(progress.seldepth),
             score: Some(score),
-            pv: Some(progress.pv.as_slice().iter().map(|m| (*m).into()).collect()),
+            pv: Some(progress.pv.clone().into_iter().map(|m| m.into()).collect()),
             time: Some(progress.stats.time),
             nodes: Some(progress.stats.nodes),
             nps: Some(progress.stats.nodes_per_second),
@@ -126,8 +126,8 @@ impl UciReporter {
         );
 
         print!("  ");
-        for mv in progress.pv.as_slice() {
-            let san_mv = san::format_move(&game, *mv);
+        for mv in progress.pv.clone().into_iter() {
+            let san_mv = san::format_move(&game, mv);
 
             print!(
                 " {}",
@@ -137,7 +137,7 @@ impl UciReporter {
                 }
             );
 
-            game.make_move(*mv);
+            game.make_move(mv);
         }
 
         println!();